# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/room": {
            "get": {
                "description": "list room entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List room entities",
                "operationId": "list-room",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Room"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create room",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create room",
                "operationId": "create-room",
                "parameters": [
                    {
                        "description": "Room entity",
                        "name": "room",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Room"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Room"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/room/{id}": {
            "get": {
                "description": "get room by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a room entity by ID",
                "operationId": "get-room",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Room ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Room"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "put": {
                "description": "update room by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a room entity by ID",
                "operationId": "update-room",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Room ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Room entity",
                        "name": "Room",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Room"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Room"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get room by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a room entity by ID",
                "operationId": "delete-room",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Room ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/roominfo": {
            "get": {
                "description": "list roominfo entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List roominfo entities",
                "operationId": "list-roominfo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.RoomInfo"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create roominfo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create roominfo",
                "operationId": "create-roominfo",
                "parameters": [
                    {
                        "description": "RoomInfo entity",
                        "name": "roominfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.RoomInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.RoomInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/roominfo/{id}": {
            "get": {
                "description": "get roominfo by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a roominfo entity by ID",
                "operationId": "get-roominfo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "RoomInfo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.RoomInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "put": {
                "description": "update roominfo by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a roominfo entity by ID",
                "operationId": "update-roominfo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "RoomInfo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "RoomInfo entity",
                        "name": "RoomInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.RoomInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.RoomInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get roominfo by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a roominfo entity by ID",
                "operationId": "delete-roominfo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "RoomInfo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ent.Book": {
            "type": "object",
            "properties": {
                "RESERVATIONS": {
                    "description": "RESERVATIONS holds the value of the \"RESERVATIONS\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the BookQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.BookEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.BookEdges": {
            "type": "object",
            "properties": {
                "bookbookstatus": {
                    "description": "BOOKBOOKSTATUS holds the value of the BOOK_BOOKSTATUS edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Bookstatus"
                },
                "bookroom": {
                    "description": "BOOKROOM holds the value of the BOOK_ROOM edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Room"
                },
                "bookuser": {
                    "description": "BOOKUSER holds the value of the BOOK_USER edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.User"
                }
            }
        },
        "ent.Bookstatus": {
            "type": "object",
            "properties": {
                "BOOKSTATUSDATA": {
                    "description": "BOOKSTATUSDATA holds the value of the \"BOOKSTATUSDATA\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the BookstatusQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.BookstatusEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.BookstatusEdges": {
            "type": "object",
            "properties": {
                "bookstatusbook": {
                    "description": "BOOKSTATUSBOOK holds the value of the BOOKSTATUS_BOOK edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Book"
                    }
                }
            }
        },
        "ent.Room": {
            "type": "object",
            "properties": {
                "ROOMNAME": {
                    "description": "ROOMNAME holds the value of the \"ROOMNAME\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RoomQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.RoomEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.RoomEdges": {
            "type": "object",
            "properties": {
                "roombook": {
                    "description": "ROOMBOOK holds the value of the ROOM_BOOK edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Book"
                    }
                },
                "roominfo": {
                    "description": "ROOMINFO holds the value of the ROOM_INFO edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.RoomInfo"
                },
                "roomroomtype": {
                    "description": "ROOMROOMTYPE holds the value of the ROOM_ROOMTYPE edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.RoomType"
                },
                "roomstatus": {
                    "description": "ROOMSTATUS holds the value of the ROOM_STATUS edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.RoomStatus"
                }
            }
        },
        "ent.RoomInfo": {
            "type": "object",
            "properties": {
                "INFOBED": {
                    "description": "INFOBED holds the value of the \"INFOBED\" field.",
                    "type": "integer"
                },
                "INFOOFFICEDE": {
                    "description": "INFOOFFICEDE holds the value of the \"INFOOFFICEDE\" field.",
                    "type": "integer"
                },
                "INFOREFRIGERAT": {
                    "description": "INFOREFRIGERAT holds the value of the \"INFOREFRIGERAT\" field.",
                    "type": "integer"
                },
                "INFOWARDROB": {
                    "description": "INFOWARDROB holds the value of the \"INFOWARDROB\" field.",
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RoomInfoQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.RoomInfoEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.RoomInfoEdges": {
            "type": "object",
            "properties": {
                "inforoom": {
                    "description": "INFOROOM holds the value of the INFO_ROOM edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Room"
                    }
                }
            }
        },
        "ent.RoomStatus": {
            "type": "object",
            "properties": {
                "STATUSDATA": {
                    "description": "STATUSDATA holds the value of the \"STATUSDATA\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RoomStatusQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.RoomStatusEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.RoomStatusEdges": {
            "type": "object",
            "properties": {
                "statusroom": {
                    "description": "STATUSROOM holds the value of the STATUS_ROOM edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Room"
                    }
                }
            }
        },
        "ent.RoomType": {
            "type": "object",
            "properties": {
                "COSTDATA": {
                    "description": "COSTDATA holds the value of the \"COSTDATA\" field.",
                    "type": "integer"
                },
                "ROOMTYPEDATA": {
                    "description": "ROOMTYPEDATA holds the value of the \"ROOMTYPEDATA\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RoomTypeQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.RoomTypeEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.RoomTypeEdges": {
            "type": "object",
            "properties": {
                "roomtyperoom": {
                    "description": "ROOMTYPEROOM holds the value of the ROOMTYPE_ROOM edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Room"
                    }
                }
            }
        },
        "ent.User": {
            "type": "object",
            "properties": {
                "USEREMAIL": {
                    "description": "USEREMAIL holds the value of the \"USEREMAIL\" field.",
                    "type": "string"
                },
                "USERNAME": {
                    "description": "USERNAME holds the value of the \"USERNAME\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the UserQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.UserEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.UserEdges": {
            "type": "object",
            "properties": {
                "userbook": {
                    "description": "USERBOOK holds the value of the USER_BOOK edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Book"
                    }
                },
                "useruserstatus": {
                    "description": "USERUSERSTATUS holds the value of the USER_USERSTATUS edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.UserStatus"
                }
            }
        },
        "ent.UserStatus": {
            "type": "object",
            "properties": {
                "STATUS": {
                    "description": "STATUS holds the value of the \"STATUS\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the UserStatusQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.UserStatusEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.UserStatusEdges": {
            "type": "object",
            "properties": {
                "userstatususer": {
                    "description": "USERSTATUSUSER holds the value of the USERSTATUS_USER edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.User"
                    }
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
...
}